Se optimizó el código; se redujo el número de archivos y código

+++ app/src/main/java/felix/alfonso/thecheezery_00000204270/dominio/Product.kt
@@ -3,4 +3,4 @@ package felix.alfonso.thecheezery_00000204270.dominio
 data class Product(var name: String,
 var image: Int,
 var description: String,
-var price: Double)
+var price: Int)
